requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
python-dotenv==1.0.1
google-genai==1.31.0
pydantic==2.8.2